+++ tasq/__init__.py
@@ -4,8 +4,8 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 
 from tasq.jobqueue import JobQueue
-from tasq.worker import ProcessWorker, ThreadWorker
+from tasq.worker import ProcessQueueWorker, ThreadQueueWorker
 from tasq.remote.client import TasqClient
-from tasq.remote.master import ActorMaster, ProcessMaster, Masters
+from tasq.remote.master import ActorMaster, QueueMaster, Masters
 
-__version__ = '1.0.0'
+__version__ = '1.0.1'